% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm.R
\name{mcm_card}
\alias{mcm_card}
\title{searc mcm for card details}
\usage{
mcm_card(x, lang = "en", ...)
}
\arguments{
\item{x}{mcm() result}

\item{lang}{language "en" or "de"}

\item{...}{for txtProgressBar}
}
\description{
searc mcm for card details
}
\examples{
\dontrun{
# card <- mcm("Avatar der Entschlossenen")
# card

val <- mcm_card(card)
val
}
}
