% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm_set.R
\name{mcm_set}
\alias{mcm_set}
\title{searc mcm for card sets}
\usage{
mcm_set(name, lang = "en")
}
\arguments{
\item{name}{uniquely identifiing set name}

\item{lang}{not yet implemented}
}
\description{
searc mcm for card sets
}
\examples{
\dontrun{
# card <- mcm_set("Rivals")
# card
}
}
