% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm_vec.R
\name{mcm_vec}
\alias{mcm_vec}
\title{search mcm for vector of cards}
\usage{
mcm_vec(vec, lang = "en", progress = TRUE)
}
\arguments{
\item{vec}{character vector of cards to look for}

\item{lang}{language en or de}
}
\description{
search mcm for vector of cards
}
\examples{
\dontrun{
cards <- mcm_vec(c("Sol Ring", "Black Lotus"))
cards
}
}
