% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm.R
\name{mcm}
\alias{mcm}
\title{search mcm for cards}
\usage{
mcm(name)
}
\arguments{
\item{name}{name of card to look for}
}
\description{
search mcm for cards
}
\examples{
\dontrun{
card <- mcm("Avatar der Entschlossenen")
card
}
}
