% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcm.R
\name{mcm}
\alias{mcm}
\title{search mcm for cards}
\usage{
mcm(card, lang = "en")
}
\arguments{
\item{lang}{language en or de}

\item{name}{card of card to look for}
}
\description{
search mcm for cards
}
\examples{
\dontrun{
card <- mcm("Black Lotus")
card

card <- mcm("Avatar der Entschlossenen", lang = "de")
card
}
}
